fix: bad size of expiration time causes read fails
was uint16_t, but needs to be `unsigned long` (uint32_t).

Also added some debug statemewnts in NexQueues

+++ include/NexQueues.h
@@ -38,7 +38,7 @@ struct nexQueuedCommand
         stringCallback strCB;
     };
     failureCallback failCB; // if we get the wrong message header
-    uint16_t expiration_time; // when should we assume it's timed out?
+    unsigned long expiration_time; // when should we assume it's timed out?
     CommandTypes cmdType; // what kind of command is it?
     nexResponses *response; // if storing response, where to?
